pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["João Tedeschi <joaorafaelbt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pytest = "^8.2.2"
pandas = "^2.2.2"
hubspot-api-client = "^9.0.0"
pyarrow = "^16.1.0"
duckdb = "^1.0.0"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "./etl",
]